two setup.sh tweaks
In Upstart scripts the $HOME environment variable may not have been set.
If so, set it now. Also, when a Virtual Machine has been corrupted, we
have to create it from scratch, but this can only happen when its
original files do not exist anymore, so make sure they've been removed.