default system32 binaries to 64-bit analysis on 64-bit machines
Now that we have proper support for 64-bit analysis (and some targets
actually provide more analysis results in 64-bit mode, such as targets
that involve Javascript) it is safe to default to 64-bit analysis for
binaries that lay in System32.

As C:\Windows\Sysnative is only present in 32-bit mode on a 64-bit
machine, as an alias, not an actual file, it is not a very good idea to
use that here. One of the issues involved is the way we use the
self.shortpath(path) trick to circumvent unicode issues when starting
new processes in Python. (As the \Sysnative is not an actual path the
shortpath function would fail on it and return a partial path).